back-end/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
Pillow==9.4.0
python-dotenv==0.20.0
openai==0.27.0